feat(system-message): add get unread msg count API

+++ model/system_message.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type SystemMessage struct {
-	ID        int64     `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID        uint      `gorm:"primaryKey;autoIncrement" json:"id"`
 	CreatedAt time.Time `gorm:"index:idx_email_created" json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 	UserEmail string    `gorm:"not null;index:idx_email_created" json:"user_email"`
